requirements.txt
accelerate==1.6.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
atlastk==0.13.4
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
databases==0.9.0
decorator==5.2.1
dotenv==0.9.9
fastapi==0.115.11
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iso8601==2.1.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
moviepy==2.1.2
mpmath==1.3.0
narwhals==1.31.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==10.4.0
proglog==0.1.11
protobuf==5.29.3
psutil==7.0.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
PyJWT==2.10.1
pypdfium2==4.30.1
pypika-tortoise==0.5.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
safetensors==0.5.3
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
SQLAlchemy==2.0.39
starlette==0.46.1
streamlit==1.44.1
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tortoise==0.1.1
tortoise-orm==0.24.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
websockets==15.0.1
